+++ Source/Mosa.Tool.Launcher/MainForm.cs
@@ -98,8 +98,8 @@ namespace Mosa.Tool.Launcher
 
 			switch (cbBootFileSystem.SelectedIndex)
 			{
-				case 0: Options.FileSystemFormat = FileSystemFormat.FAT16; ; break;
-				case 1: Options.FileSystemFormat = FileSystemFormat.FAT24; ; break;
+				case 0: Options.FileSystemFormat = FileSystemFormat.FAT12; ; break;
+				case 1: Options.FileSystemFormat = FileSystemFormat.FAT16; ; break;
 				default: Options.FileSystemFormat = FileSystemFormat.FAT16; ; break;
 			}
 
@@ -157,8 +157,8 @@ namespace Mosa.Tool.Launcher
 
 			switch (Options.FileSystemFormat)
 			{
-				case FileSystemFormat.FAT16: cbBootFileSystem.SelectedIndex = 0; break;
-				case FileSystemFormat.FAT24: cbBootFileSystem.SelectedIndex = 1; break;
+				case FileSystemFormat.FAT12: cbBootFileSystem.SelectedIndex = 0; break;
+				case FileSystemFormat.FAT16: cbBootFileSystem.SelectedIndex = 1; break;
 				default: break;
 			}
